Refactors Shopify initialization into hook
